Fix bug in truncated host elements
For example, "[::1", where the square bracket at
the end is missing.

Handle truncated host element by checking the
state we end up in after the entire string is parsed.
Some states represent interal elements of a host name
or address, and so if we're still in those states and
run out of input characters, the input string was
cut off early.